feat: slide out pantry drawer!
