+++ frontend/src/components/PantryCard/index.ts
@@ -1,0 +1,1 @@
+export { default } from './PantryCard';
